% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildApplicationFile}
\alias{buildApplicationFile}
\title{Prepares a new network visualization}
\usage{
buildApplicationFile(mtdt, path_to_file, network.name,
  q.threshold = 0.01, mutation_effects = NULL)
}
\arguments{
\item{mtdt}{An indicator matrix, of the same type used in the DISCOVER test. The dimensions should be n genes (rows) by m columns (samples), with cells being 0 (non-mutated) or 1 (mutated).}

\item{path_to_file}{Where to save the results. This file can be uploaded to the associated application.}

\item{network.name}{The name of the network that will be displayed within the application.}

\item{q.threshold}{The FDR cutoff used by the discover test, defaults to 0.01.}

\item{mutation_effects}{Supply a list of mutations, with a column "gene" and "functionalEffect" to include mutation effects in the visualization. This is not necessary for the majority of functionality, therefore this function will run if it is not supplied.}
}
\description{
This is the main function of the package, meant to prepare mutation data to be visualized by the discoverInteractiveAnalytics application.
}
