% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{groupTestByGeneOntology}
\alias{groupTestByGeneOntology}
\title{Run DISCOVER group-wise tests on gene sets defined by gene ontology terms}
\usage{
groupTestByGeneOntology(events, dictionary = NULL, go.type = "bp",
  group.test.type = "exclusivity")
}
\arguments{
\item{events}{Can either be a discover matrix that results from running the discover.pairwise.test() or an indicator matrix. If it is an indicator matrix, the function will run the discover test, but this will significantly increase execution time.}

\item{dictionary}{If a dictionary object (\code{\link{makeDictionary}}) is not supplied this function will automatically generate one, though this increases execution time.}

\item{go.type}{Character. Can be "bp","cc" or "mf" for biological process, cellular component, or molecular function respectively. Defaults to "bp"}

\item{group.test.type}{The type of group test to be employed: "exclusivity", "coverage" or "impurity"}
}
\description{
This implements the discover group wise tests of exclusivity, coverage and impurity \link{"https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-1114-x"}.
Automatically implements a the Benjamini-Hochberg multiple testing correction.
}
