% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{makeDictionary}
\alias{makeDictionary}
\title{Make a dictionary of gene ontology terms, given a list of genes.}
\usage{
makeDictionary(events)
}
\arguments{
\item{mtdt}{Can either be an indicator matrix of the type used by the discover test (see ?\code{\link{buildDirectory}}) or a list of genes.}
}
\value{
This will return a nested list. The list can be searched by either genes (searching associated gene ontology terms) or by gene ontology term (in this case searching for associated genes).
}
\description{
Uses myGene.io to build a dictionary of gene ontology terms. It relies on myGene's query many function to build this list.
The difference here is this function also allows for forward and reverse lookups (either go terms associated with a gene, or genes assoicated with a go term).
This is largely to improve performance within the application. This function is also necessary to run group-wise tests defined by gene ontology terms.
}
