% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{testEnrichment}
\alias{testEnrichment}
\title{Test for Enrichment of particular gene ontology terms}
\usage{
testEnrichment(res.mutex, dictionary, gene_ontology_domain = c("bp",
  "cc", "mf"), resamples = 1000)
}
\description{
Test for Enrichment of particular gene ontology terms
}
